Add files via upload
code updated to reflect below requirements

"I had a button there to use for the filter element (the little filter/funnel icon button that is grayed out). You could either use that existing button or get rid of it so we don't have redundant stuff. The same button is present on the listing section of the dashboard page, where your filtering logic should also be dropped in
We might get a poor UX here by doing this solely on the client side - for example if the closest 100 listings were not the charger type the user is looking for, they would get no results, when they might actually have results within a usable distance. You could add some logic to do another query client side to get more results and try filtering again (e.g. change the result count to 300), or update the server function so it takes the filter parameter and just get the filtered results on the first call. This is an edge case, i dont think it's critical that it's handled for this iteration, but it is worth dealing with if you have the chance to"